# Support late materialization in the Parquet reader, and handle `COUNT(*)` directly in the multi file reader