--- a/src/include/duckdb/common/multi_file/multi_file_reader.hpp
+++ b/src/include/duckdb/common/multi_file/multi_file_reader.hpp
@@ -241,6 +241,8 @@ public:
 	                           const LogicalType &type, MultiFileLocalIndex local_index,
 	                           optional_ptr<MultiFileColumnDefinition> &global_column_reference);
 
+	DUCKDB_API virtual unique_ptr<MultiFileReader> Copy() const;
+
 protected:
 	//! Used in errors to report which function is using this MultiFileReader
 	string function_name;