--- a/src/include/duckdb/function/table_function.hpp
+++ b/src/include/duckdb/function/table_function.hpp
@@ -318,6 +318,9 @@ typedef vector<PartitionStatistics> (*table_function_get_partition_stats_t)(Clie
 typedef virtual_column_map_t (*table_function_get_virtual_columns_t)(ClientContext &context,
                                                                      optional_ptr<FunctionData> bind_data);
 
+typedef vector<column_t> (*table_function_get_row_id_columns)(ClientContext &context,
+                                                              optional_ptr<FunctionData> bind_data);
+
 //! When to call init_global to initialize the table function
 enum class TableFunctionInitialization { INITIALIZE_ON_EXECUTE, INITIALIZE_ON_SCHEDULE };
 
@@ -396,6 +399,8 @@ public:
 	table_function_get_partition_stats_t get_partition_stats;
 	//! (Optional) returns a list of virtual columns emitted by the table function
 	table_function_get_virtual_columns_t get_virtual_columns;
+	//! (Optional) returns a list of row id columns
+	table_function_get_row_id_columns get_row_id_columns;
 
 	table_function_serialize_t serialize;
 	table_function_deserialize_t deserialize;