--- a/src/include/duckdb/optimizer/late_materialization.hpp
+++ b/src/include/duckdb/optimizer/late_materialization.hpp
@@ -28,8 +28,8 @@ private:
 	bool TryLateMaterialization(unique_ptr<LogicalOperator> &op);
 
 	unique_ptr<LogicalGet> ConstructLHS(LogicalGet &get);
-	ColumnBinding ConstructRHS(unique_ptr<LogicalOperator> &op);
-	idx_t GetOrInsertRowId(LogicalGet &get);
+	vector<ColumnBinding> ConstructRHS(unique_ptr<LogicalOperator> &op);
+	vector<idx_t> GetOrInsertRowIds(LogicalGet &get);
 
 	void ReplaceTopLevelTableIndex(LogicalOperator &op, idx_t new_index);
 	void ReplaceTableReferences(Expression &expr, idx_t new_table_index);
@@ -41,8 +41,10 @@ private:
 	Optimizer &optimizer;
 	//! The max row count for which we will consider late materialization
 	idx_t max_row_count;
-	//! The type of the row id column
-	LogicalType row_id_type;
+	//! The row-id column ids
+	vector<column_t> row_id_column_ids;
+	//! The row-id columns
+	vector<TableColumn> row_id_columns;
 };
 
 } // namespace duckdb