--- a/src/include/duckdb/planner/filter/conjunction_filter.hpp
+++ b/src/include/duckdb/planner/filter/conjunction_filter.hpp
@@ -36,7 +36,7 @@ public:
 
 public:
 	ConjunctionOrFilter();
-	FilterPropagateResult CheckStatistics(BaseStatistics &stats) override;
+	FilterPropagateResult CheckStatistics(BaseStatistics &stats) const override;
 	string ToString(const string &column_name) const override;
 	bool Equals(const TableFilter &other) const override;
 	unique_ptr<TableFilter> Copy() const override;
@@ -53,7 +53,7 @@ public:
 	ConjunctionAndFilter();
 
 public:
-	FilterPropagateResult CheckStatistics(BaseStatistics &stats) override;
+	FilterPropagateResult CheckStatistics(BaseStatistics &stats) const override;
 	string ToString(const string &column_name) const override;
 	bool Equals(const TableFilter &other) const override;
 	unique_ptr<TableFilter> Copy() const override;