--- a/src/include/duckdb/planner/filter/constant_filter.hpp
+++ b/src/include/duckdb/planner/filter/constant_filter.hpp
@@ -28,7 +28,7 @@ public:
 
 public:
 	bool Compare(const Value &value) const;
-	FilterPropagateResult CheckStatistics(BaseStatistics &stats) override;
+	FilterPropagateResult CheckStatistics(BaseStatistics &stats) const override;
 	string ToString(const string &column_name) const override;
 	bool Equals(const TableFilter &other) const override;
 	unique_ptr<TableFilter> Copy() const override;