--- a/src/include/duckdb/planner/filter/dynamic_filter.hpp
+++ b/src/include/duckdb/planner/filter/dynamic_filter.hpp
@@ -37,7 +37,7 @@ public:
 	shared_ptr<DynamicFilterData> filter_data;
 
 public:
-	FilterPropagateResult CheckStatistics(BaseStatistics &stats) override;
+	FilterPropagateResult CheckStatistics(BaseStatistics &stats) const override;
 	string ToString(const string &column_name) const override;
 	bool Equals(const TableFilter &other) const override;
 	unique_ptr<TableFilter> Copy() const override;