--- a/src/include/duckdb/planner/filter/in_filter.hpp
+++ b/src/include/duckdb/planner/filter/in_filter.hpp
@@ -23,7 +23,7 @@ public:
 	vector<Value> values;
 
 public:
-	FilterPropagateResult CheckStatistics(BaseStatistics &stats) override;
+	FilterPropagateResult CheckStatistics(BaseStatistics &stats) const override;
 	string ToString(const string &column_name) const override;
 	bool Equals(const TableFilter &other) const override;
 	unique_ptr<TableFilter> Copy() const override;