--- a/src/include/duckdb/planner/filter/optional_filter.hpp
+++ b/src/include/duckdb/planner/filter/optional_filter.hpp
@@ -27,7 +27,7 @@ public:
 	string ToString(const string &column_name) const override;
 	unique_ptr<TableFilter> Copy() const override;
 	unique_ptr<Expression> ToExpression(const Expression &column) const override;
-	FilterPropagateResult CheckStatistics(BaseStatistics &stats) override;
+	FilterPropagateResult CheckStatistics(BaseStatistics &stats) const override;
 	void Serialize(Serializer &serializer) const override;
 	static unique_ptr<TableFilter> Deserialize(Deserializer &deserializer);
 };