--- a/src/include/duckdb/planner/filter/struct_filter.hpp
+++ b/src/include/duckdb/planner/filter/struct_filter.hpp
@@ -31,7 +31,7 @@ public:
 	unique_ptr<TableFilter> child_filter;
 
 public:
-	FilterPropagateResult CheckStatistics(BaseStatistics &stats) override;
+	FilterPropagateResult CheckStatistics(BaseStatistics &stats) const override;
 	string ToString(const string &column_name) const override;
 	bool Equals(const TableFilter &other) const override;
 	unique_ptr<TableFilter> Copy() const override;