--- a/src/include/duckdb/planner/table_filter.hpp
+++ b/src/include/duckdb/planner/table_filter.hpp
@@ -48,7 +48,7 @@ public:
 
 public:
 	//! Returns true if the statistics indicate that the segment can contain values that satisfy that filter
-	virtual FilterPropagateResult CheckStatistics(BaseStatistics &stats) = 0;
+	virtual FilterPropagateResult CheckStatistics(BaseStatistics &stats) const = 0;
 	virtual string ToString(const string &column_name) const = 0;
 	string DebugToString() const;
 	virtual unique_ptr<TableFilter> Copy() const = 0;