… benchmark/tpch/parquet/parquet_top_n.benchmark
… extension/parquet/CMakeLists.txt
… extension/parquet/parquet_extension.cpp
… src/optimizer/late_materialization.cpp
… src/planner/filter/struct_filter.cpp
… src/storage/statistics/numeric_stats.cpp
… src/storage/statistics/string_stats.cpp
… src/storage/table/row_group.cpp
… test/sql/copy/parquet/batched_write/tpch_sf1_parquet.test_slow
… test/sql/copy/parquet/parquet_count_star.test
… test/sql/copy/parquet/parquet_late_materialization.test