This PR generalizes the late materialization optimizer introduced in https://github.com/duckdb/duckdb/pull/15692 - allowing it to be used for the Parquet reader.

In particular, the `TableFunction` is extended with an extra callback that allows specifying the relevant row-id columns:

```cpp
typedef vector<column_t> (*table_function_get_row_id_columns)(ClientContext &context,
                                                              optional_ptr<FunctionData> bind_data);
```

This is then used by the Parquet reader to specify the two row-id columns: `file_index` (https://github.com/duckdb/duckdb/pull/17144) and `file_row_number` (https://github.com/duckdb/duckdb/pull/16979). Top-N , sample and limit/offset queries are then transformed into a join on the relevant row-id columns. For example:


```sql
SELECT * FROM lineitem.parquet ORDER BY l_extendedprice DESC LIMIT 5;

-- becomes

SELECT * FROM lineitem.parquet WHERE (file_index, file_row_number) IN (
    SELECT file_index, file_row_number FROM lineitem.parquet ORDER BY l_extendedprice DESC LIMIT 5)
ORDER BY l_extendedprice DESC;
```

### Performance

```sql
SELECT * FROM lineitem.parquet ORDER BY l_extendedprice DESC LIMIT 5;
```

| v1.2.1 |  main  |  new   |
|--------|--------|--------|
| 0.19s  | 0.14s | 0.06s |


```sql
SELECT * FROM lineitem.parquet ORDER BY l_orderkey DESC LIMIT 5;
```
| v1.2.1 | main  |  new  |
|--------|-------|-------|
| 0.73s  | 0.53s | 0.06s |

```sql
SELECT * FROM lineitem.parquet LIMIT 1000000 OFFSET 10000000;
```
| v1.2.1 | main  |  new  |
|--------|-------|-------|
| 1.6s   | 1.2s | 0.14s |


### Refactor

I've also moved the `ParquetMultiFileInfo` to a separate file as part of this PR - which is most of the changes here.