@djouallah 2025-05-03
@Mytherin do you have any plan to extend this to filtering ?  for example 

`from './orders/*.parquet'  where o_orderkey = 1`